backend/requirements.txt
fastapi[all]
uvicorn[standard]
pydantic[email]
motor
bcrypt
python-jose
black
flake8